add user validation in blogs

+++ controllers/blogs.js
@@ -13,14 +13,10 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
+  const user = request.user
 
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if(!decodedToken.id){
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
   if(!user){
-    return response.status(400).json({ error: 'user not found' })
+    return response.status(401).json({error: 'unauthorized'})
   }
 
   const blog = new Blog({
@@ -47,6 +43,14 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+  if(!user){
+    return response.status(401).json({error: 'unauthorized'})
+  }
+  if(!(user.id.toString() === blog.user.id.toString())){
+    return response.status(401).json({ error: 'unauthorized delete' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
